Cleaned up BackgroundLayer. Moved Cursors and Viewports into their own seperate components. Removed hooks from within the render functions.
