+++ webstack/apps/webapp/src/app/components/Board/BackgroundLayer.tsx
@@ -6,31 +6,17 @@
  *
  */
 
-import { useCallback, useEffect, useState, Fragment } from 'react';
-import { Box, Tag } from '@chakra-ui/react';
-import { motion, useAnimation } from 'framer-motion';
+import { useEffect, useState } from 'react';
 
 import { DraggableEvent } from 'react-draggable';
-import { ErrorBoundary } from 'react-error-boundary';
-import { GiArrowCursor } from 'react-icons/gi';
 import { DraggableData, Rnd } from 'react-rnd';
-import { throttle } from 'throttle-debounce';
 
-import {
-  useAppStore,
-  useHexColor,
-  useHotkeys,
-  usePresence,
-  usePresenceStore,
-  useUIStore,
-  useUser,
-  useUsersStore,
-  useWindowResize,
-} from '@sage3/frontend';
-import { Applications, AppError } from '@sage3/applications/apps';
+import { useAppStore, useUIStore } from '@sage3/frontend';
 
 import { Background } from './Background/Background';
-import { AppWindow } from '@sage3/applications/apps';
+import { Apps } from './Background/Apps';
+import { Cursors } from './Background/Cursors';
+import { Viewports } from './Background/Viewports';
 
 type BackgroundLayerProps = {
   boardId: string;
@@ -38,9 +24,6 @@ type BackgroundLayerProps = {
 };
 
 export function BackgroundLayer(props: BackgroundLayerProps) {
-  // User
-  const { user } = useUser();
-
   // Apps Store
   const apps = useAppStore((state) => state.apps);
   const appsFetched = useAppStore((state) => state.fetched);
@@ -53,20 +36,13 @@ export function BackgroundLayer(props: BackgroundLayerProps) {
   const setBoardPosition = useUIStore((state) => state.setBoardPosition);
   const setScale = useUIStore((state) => state.setScale);
   const boardPosition = useUIStore((state) => state.boardPosition);
-  const resetZIndex = useUIStore((state) => state.resetZIndex);
   const boardDragging = useUIStore((state) => state.boardDragging);
   const setBoardDragging = useUIStore((state) => state.setBoardDragging);
   const fitApps = useUIStore((state) => state.fitApps);
   const boardLocked = useUIStore((state) => state.boardLocked);
 
-  // Presence Information
-  const { update: updatePresence } = usePresence();
-  const presences = usePresenceStore((state) => state.presences);
-  const users = useUsersStore((state) => state.users);
-
   // Local State
   const [boardDrag, setBoardDrag] = useState(false); // Used to differentiate between board drag and app deselect
-  const { width: winWidth, height: winHeight } = useWindowResize();
 
   // Position board when entering board
   useEffect(() => {
@@ -106,71 +82,6 @@ export function BackgroundLayer(props: BackgroundLayerProps) {
     setBoardDrag(false);
   }
 
-  // Reset the global zIndex when no apps
-  useEffect(() => {
-    if (apps.length === 0) resetZIndex();
-  }, [apps]);
-
-  // Update the user's cursor every half second
-  const throttleCursor = throttle(500, (e: MouseEvent) => {
-    if (boardDragging) return;
-    const winX = e.clientX;
-    const winY = e.clientY;
-    const bx = boardPosition.x;
-    const by = boardPosition.y;
-    const s = scale;
-    const x = winX / s - bx;
-    const y = winY / s - by;
-    const z = 0;
-    updatePresence({ cursor: { x, y, z } });
-  });
-
-  // Keep a copy of the function
-  const throttleCursorFunc = useCallback(throttleCursor, [boardPosition.x, boardPosition.y, scale, boardDragging]);
-  const cursorFunc = (e: MouseEvent) => {
-    // Check if event is on the board
-    if (updatePresence) {
-      // Send the throttled version to the server
-      throttleCursorFunc(e);
-    }
-  };
-
-  // Attach the mouse move event to the window
-  useEffect(() => {
-    const mouseMove = (e: MouseEvent) => cursorFunc(e);
-    window.addEventListener('mousemove', mouseMove);
-    return () => window.removeEventListener('mousemove', mouseMove);
-  }, [boardPosition.x, boardPosition.y, scale, boardDragging]);
-
-  // Update the user's viewport every half second
-  const throttleViewport = throttle(500, (x: number, y: number, width: number, height: number) => {
-    const viewPos = { x, y, z: 0 };
-    const viewWidth = width;
-    const viewHeight = height;
-    const viewSize = { width: viewWidth, height: viewHeight };
-    updatePresence({ viewport: { position: viewPos, size: viewSize } });
-  });
-
-  // Keep a copy of the function
-  const throttleViewportFunc = useCallback(throttleViewport, []);
-  const viewportFunc = (x: number, y: number, w: number, h: number) => {
-    // Check if event is on the board
-    if (updatePresence) {
-      // Send the throttled version to the server
-      throttleViewportFunc(x, y, w, h);
-    }
-  };
-
-  // Update Viewport Presence
-  useEffect(() => {
-    viewportFunc(-boardPosition.x, -boardPosition.y, winWidth / scale, winHeight / scale);
-  }, [boardPosition.x, boardPosition.y, winWidth, winHeight, scale]);
-
-  // Deselect all apps
-  useHotkeys('esc', () => {
-    setSelectedApp('');
-  });
-
   return (
     <div style={{ transform: `scale(${scale})`, transformOrigin: 'top left' }}>
       {/* Board. Uses lib react-rnd for drag events.
@@ -192,147 +103,18 @@ export function BackgroundLayer(props: BackgroundLayerProps) {
         dragHandleClassName={'board-handle'}
         disableDragging={boardLocked}
       >
-        {/* Apps */}
-        {apps.map((app) => {
-          if (app.data.type in Applications) {
-            const Component = Applications[app.data.type].AppComponent;
-            return (
-              // Wrap the components in an errorboundary to protect the board from individual app errors
-              <ErrorBoundary
-                key={app._id}
-                fallbackRender={({ error, resetErrorBoundary }) => (
-                  <AppError error={error} resetErrorBoundary={resetErrorBoundary} app={app} />
-                )}
-              >
-                <Component key={app._id} {...app}></Component>
-              </ErrorBoundary>
-            );
-          } else {
-            // App not found: happens if unkonw app in sagebase
-            return (
-              <AppWindow key={app._id} app={app}>
-                <div>App not found</div>
-              </AppWindow>
-            );
-          }
-        })}
+        {/* The board's apps */}
+        <Apps />
 
-        {/* Draw the cursors and viewports: filter by board and not myself */}
-        {presences
-          .filter((el) => el.data.boardId === props.boardId)
-          .filter((el) => el.data.userId !== user?._id)
-          .map((presence) => {
-            const u = users.find((el) => el._id === presence.data.userId);
-            if (!u) return null;
-            const color = useHexColor(u.data.color || 'red');
-            const isWall = u.data.userType === 'wall';
-            if (isWall) {
-              return <Fragment key={u._id}>
-                <Box
-                  borderStyle="dashed"
-                  borderWidth={3 / scale}
-                  borderColor={color}
-                  borderTop={'none'}
-                  position="absolute"
-                  pointerEvents="none"
-                  left={presence.data.viewport.position.x + 'px'}
-                  top={presence.data.viewport.position.y + 'px'}
-                  width={presence.data.viewport.size.width + 'px'}
-                  height={presence.data.viewport.size.height + 'px'}
-                  opacity={0.8}
-                  borderRadius="8px 8px 8px 8px"
-                  transition="all 0.5s"
-                  color="white"
-                  fontSize="xl"
-                  background={`linear-gradient(180deg, ${color} 30px, transparent 30px)`}
-                >
-                  Viewport for {u.data.name}
-                </Box>
-                <UserCursor
-                  color={color}
-                  position={presence.data.cursor}
-                  name={users.find((el) => el._id === presence.data.userId)?.data.name || '-'}
-                  scale={scale}
-                />
-              </Fragment>
-            } else {
-              return (
-                <UserCursor
-                  key={'user' + u._id}
-                  color={color}
-                  position={presence.data.cursor}
-                  name={users.find((el) => el._id === presence.data.userId)?.data.name || '-'}
-                  scale={scale}
-                />
-              );
-            }
-          })}
+        {/* User Cursors */}
+        <Cursors boardId={props.boardId} />
+
+        {/* User Viewports */}
+        <Viewports boardId={props.boardId} />
 
         {/* Draggable Background */}
         <Background boardId={props.boardId} roomId={props.roomId}></Background>
       </Rnd>
     </div>
-  );
-}
-
-/**
- * User Curor Props
- */
-interface UserCursorProps {
-  name: string;
-  color: string;
-  position: { x: number; y: number; z: number };
-  scale: number;
-}
-
-/**
- * Show a user pointer
- * @param props UserCursorProps
- * @returns
- */
-function UserCursor(props: UserCursorProps) {
-  // Create an animation object to control the opacity of pointer
-  // Pointer fades after inactivity
-  const controls = useAnimation();
-
-  // Reset animation if pointer moves
-  useEffect(() => {
-    // Stop previous animation
-    controls.stop();
-    // Set initial opacity
-    controls.set({ opacity: 1.0 });
-    // Start animation
-    controls.start({
-      // final opacity
-      opacity: 0.0,
-      transition: {
-        ease: 'easeIn',
-        // duration in sec.
-        duration: 10,
-        delay: 30,
-      },
-    });
-  }, [props.position.x, props.position.y, props.position.z]);
-
-  return (
-    <motion.div
-      // pass the animation controller
-      animate={controls}
-      style={{
-        position: 'absolute',
-        left: props.position.x - 4 + 'px',
-        top: props.position.y - 3 + 'px',
-        transition: 'left 0.5s ease-in-out, top 0.5s ease-in-out',
-        pointerEvents: 'none',
-        display: 'flex',
-        zIndex: 100000,
-        transform: `scale(${1 / props.scale})`,
-      }}
-    >
-      <GiArrowCursor color={props.color}></GiArrowCursor>
-      <Tag variant="solid" borderRadius="md" mt="3" mb="0" ml="-1" mr="0" p="1" color="white">
-        {props.name}
-      </Tag>
-    </motion.div>
   );
 }
